ctf-agent/requirements.txt
# Core
requests>=2.31.0
pyyaml>=6.0
pydantic>=2.0

# Cryptography / math
sympy>=1.12

# Binary analysis (Python-side)
pwntools>=4.12.0

# Image processing
Pillow>=10.0.0

# Optional: symbolic execution
# angr>=9.2.0   # large install — uncomment if needed

# Optional: fuzzing Python harness
# python-afl    # requires AFL++ to be installed

# Dev / testing
pytest>=7.4.0
